backend/server/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
rapidfuzz
python-multipart
s2sphere
